Clean up redundant arguments and fix debug mode
Currently the script looks for two unique arguments to trigger, there isn't a purpose to require that. I've left the `-batchmodebuilder` since it is specific to this project.

There was a typo on the Debug flag, I fixed that. Also, I exit the editor when the arguments fail to build, rather than leaving the process hanging.

Lastly, I updated the comment as it had needless arguments (`-build` and the last `/dev/stdout`) and made it more coherent with the logfile path + tee command.